Move Text to TextElement so there is no same naming schema
