+++ frontend/src/lib/components/elements/TextElement/index.tsx
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { default } from "./Text"
+export { default } from "./TextElement"
